Add files via upload
Remote Procedure Call - Distributed System

Modify the program described in Assignment 3 in such a way that the two programs can be run on different machines using Remote Procedure Calls (RPC). Remote procedure calls allow programs to make procedure calls into remote programs running on a different machine across the network and are therefore fundamental in building a distributed system. The  receiver.c and the processor.c have the same functionality as described in Assignment 3 except that they should use remote procedure calls to communicate and achieve the same functionality. 

Please submit a zip file labeled as Assignment4_Yourname.zip. It should include a PDF file with explanation of what your code does (this is optional if your code is self explanatory with comments) and the relevant source files and a Makefile to compile the different components.